Preconfigured pages are now read only
Since the query string was the highest priority for initializing the
form, that meant that reloading the page always silently wiped your own
configuration away. That didn't seem right, so I wanted to make it more
explicit that a preconfigured page is set by the query string and
nothing else. Plus, this means that you can't accidentally reconfigure
the page after opening a preconfigured page.